@@ -10,20 +10,19 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "12"
+    "node": "18"
   },
   "main": "index.js",
   "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "firebase-admin": "^9.2.0",
-    "firebase-functions": "^3.11.0",
-    "stripe": "^8.142.0"
+    "express": "^4.18.2",
+    "firebase-admin": "^11.8.0",
+    "firebase-functions": "^4.3.1",
+    "stripe": "^14.18.0"
   },
   "devDependencies": {
-    "eslint": "^7.6.0",
+    "eslint": "^8.15.0",
     "eslint-config-google": "^0.14.0",
-    "firebase-functions-test": "^0.2.0"
+    "firebase-functions-test": "^3.1.0"
   },
   "private": true
 }
