@@ -1,14 +1,13 @@
-import React from "react";
+import React, { useHistory } from "react";
 import "./Subtotal.css";
 import CurrencyFormat from "react-currency-format";
 import { useStateValue } from "../../StateProvider";
 import { getBasketTotal } from "../../reducer";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 function Subtotal() {
-  const history = useHistory();
+  const navigate = useNavigate();
   const [{ basket }, dispatch] = useStateValue();
-
   return (
     <div className="subtotal">
       <CurrencyFormat
@@ -17,20 +16,19 @@ function Subtotal() {
             <p>
               Subtotal ({basket.length} items): <strong>{value}</strong>
             </p>
-            <small className="subtotal-gift">
+            <small className="subtotal_gift">
               <input type="checkbox" /> This order contains a gift
             </small>
           </>
         )}
         decimalScale={2}
-        value={getBasketTotal(basket)}
+        value={getBasketTotal(basket)} // Part of the homework
         displayType={"text"}
         thousandSeparator={true}
         prefix={"$"}
       />
-      <button onClick={(e) => history.push("/payment")}>
-        Proceed to Checkout
-      </button>
+
+      <button onClick={(e) => navigate("/payment")}>Proceed to Checkout</button>
     </div>
   );
 }
