@@ -1,14 +1,28 @@
 module.exports = {
-  root: true,
   env: {
     es6: true,
     node: true,
+  },
+  parserOptions: {
+    "ecmaVersion": 2018,
   },
   extends: [
     "eslint:recommended",
     "google",
   ],
   rules: {
-    quotes: ["error", "double"],
+    "no-restricted-globals": ["error", "name", "length"],
+    "prefer-arrow-callback": "error",
+    "quotes": ["error", "double", {"allowTemplateLiterals": true}],
   },
+  overrides: [
+    {
+      files: ["**/*.spec.*"],
+      env: {
+        mocha: true,
+      },
+      rules: {},
+    },
+  ],
+  globals: {},
 };
