@@ -1,15 +1,18 @@
-const functions = require("firebase-functions");
+const { onRequest } = require("firebase-functions/v2/https");
+const logger = require("firebase-functions/logger");
+
 const express = require("express");
+
 const cors = require("cors");
+
 const stripe = require("stripe")(
-  "sk_test_51IdVudBDYzH08iN6Eay86bOJHF3Rcnk3Op924xYPxQCaCDWJR5JHRhtUYvSSC9v76vK5Cozy97kO9GPge4HKJJO9003LEFLBdA"
+  "sk_test_51OmheRCgpVnhkvFBIu1AvQ3br2b84YXQHo9C5loFpcqOhnWQBa4rCUGQbFHcIsfOY0nZK0NH17mQKajTVQuTJhGq00yCvciJMy"
 );
 
 // API
 
 // - App config
 const app = express();
-
 // - Middlewares
 app.use(cors({ origin: true }));
 app.use(express.json());
@@ -20,7 +23,7 @@ app.get("/", (request, response) => response.status(200).send("hello world"));
 app.post("/payments/create", async (request, response) => {
   const total = request.query.total;
 
-  console.log("Payment Request Recieved BOOM!!! for this amount >>> ", total);
+  logger.info("Payment Request Recieved BOOM!!! for this amount >>> ", total);
 
   const paymentIntent = await stripe.paymentIntents.create({
     amount: total, // subunits of the currency
@@ -32,9 +35,7 @@ app.post("/payments/create", async (request, response) => {
     clientSecret: paymentIntent.client_secret,
   });
 });
-
 // - Listen command
-exports.api = functions.https.onRequest(app);
+exports.api = onRequest(app);
 
-// Example endpoint
-// http://localhost:5001/amaznclone-v1/us-central1/api.
+// http://localhost:5001/clone-8ecbd/us-central1/api
