@@ -1,10 +1,10 @@
 import React, { createContext, useContext, useReducer } from "react";
 
-// Prepares the dataLayer
+// Prepare the datalayer
 export const StateContext = createContext();
 
 // Wrap our app and provide the Data layer
-export const StateProvider = ({ reducer, initialState, children }) => (
+export const StateProvider = ({ initialState, reducer, children }) => (
   <StateContext.Provider value={useReducer(reducer, initialState)}>
     {children}
   </StateContext.Provider>
