@@ -1,24 +1,23 @@
 import React from "react";
-import { useStateValue } from "../../StateProvider";
 import "./Checkout.css";
-import CheckoutProduct from "./CheckoutProduct";
-import Subtotal from "./Subtotal";
+import Subtotal from "../Subtotal/Subtotal";
+import CheckoutProduct from "../CheckoutProduct/CheckoutProduct";
+import { useStateValue } from "../../StateProvider";
 
 function Checkout() {
   const [{ basket, user }, dispatch] = useStateValue();
 
   return (
     <div className="checkout">
-      <div className="checkout-left">
+      <div className="checkout_left">
         <img
-          className="checkout-ad"
+          className="checkout_ad"
           src="https://images-na.ssl-images-amazon.com/images/G/02/UK_CCMP/TM/OCC_Amazon1._CB423492668_.jpg"
           alt=""
         />
-
         <div>
           <h3>Hello, {user?.email}</h3>
-          <h2 className="checkout-title">Your shopping Basket</h2>
+          <h2 className="checkout_title">Your shopping Basket</h2>
 
           {basket.map((item) => (
             <CheckoutProduct
@@ -32,7 +31,7 @@ function Checkout() {
         </div>
       </div>
 
-      <div className="checkout-right">
+      <div className="checkout_right">
         <Subtotal />
       </div>
     </div>
