@@ -1,30 +1,32 @@
 import React from "react";
 import "./Order.css";
 import moment from "moment";
+import CheckoutProduct from "../CheckoutProduct/CheckoutProduct";
 import CurrencyFormat from "react-currency-format";
-import CheckoutProduct from "../Checkout/CheckoutProduct";
 
 function Order({ order }) {
   return (
     <div className="order">
       <h2>Order</h2>
       <p>{moment.unix(order.data.created).format("MMMM Do YYYY, h:mma")}</p>
-      <p className="order-id">
+      <p className="order_id">
         <small>{order.id}</small>
       </p>
-      {order.data.basket?.map((item) => (
-        <CheckoutProduct
-          id={item.id}
-          title={item.title}
-          image={item.image}
-          price={item.price}
-          rating={item.rating}
-          hideButton
-        />
-      ))}
+      {order.data.basket?.map((item) => {
+        return (
+          <CheckoutProduct
+            id={item.id}
+            title={item.title}
+            image={item.image}
+            price={item.price}
+            rating={item.rating}
+            hideButton
+          />
+        );
+      })}
       <CurrencyFormat
         renderText={(value) => (
-          <h3 className="order-total">Order Total: {value}</h3>
+          <h3 className="order_total">Order Total: {value}</h3>
         )}
         decimalScale={2}
         value={order.data.amount / 100}
