@@ -1,53 +1,55 @@
-import React, { useState, useEffect } from "react";
-import { useStateValue } from "../../StateProvider";
-import CheckoutProduct from "../Checkout/CheckoutProduct";
+import React, { useEffect, useState } from "react";
 import "./Payment.css";
-import { Link, useHistory } from "react-router-dom";
-import { CardElement, useStripe, useElements } from "@stripe/react-stripe-js";
 import CurrencyFormat from "react-currency-format";
-import { getBasketTotal } from "./../../reducer";
-import axios from "./../../axios";
+import { useStateValue } from "../../StateProvider";
+import { getBasketTotal } from "../../reducer";
+import CheckoutProduct from "../CheckoutProduct/CheckoutProduct";
+import { Link, Navigate, useNavigate } from "react-router-dom";
+import { useStripe, useElements, CardElement } from "@stripe/react-stripe-js";
+import axios from "../../axios";
+// import { set } from "sanity";
 import { db } from "../../firebase";
 
 function Payment() {
   const [{ basket, user }, dispatch] = useStateValue();
-  const history = useHistory();
 
+  const [error, setError] = useState(null);
+  const [disabled, setDisabled] = useState(true);
+  const [succeeded, setSucceeded] = useState(false);
+  const [processing, setProcessing] = useState("");
+  const [clientSecret, setClientSecret] = useState(true);
+
+  const Navigate = useNavigate();
   const stripe = useStripe();
   const elements = useElements();
 
-  const [succeeded, setSucceeded] = useState(false);
-  const [processing, setProcessing] = useState("");
-  const [error, setError] = useState(null);
-  const [disabled, setDisabled] = useState(true);
-  const [clientSecret, setClientSecret] = useState(true);
-
   useEffect(() => {
     // generate the special stripe secret which allows us to charge a customer
+
     const getClientSecret = async () => {
       const response = await axios({
         method: "post",
         // Stripe expects the total in a currencies subunits
         url: `/payments/create?total=${getBasketTotal(basket) * 100}`,
       });
-      setClientSecret(response.data.clientSecret);
+      setClientSecret(response.data.clientSecret); // clientSecret is the payment intent
     };
-
     getClientSecret();
   }, [basket]);
 
   console.log("THE SECRET IS >>>", clientSecret);
-  console.log("👱", user);
 
-  const handleSubmit = async (event) => {
+  const handleSubmit = async (e) => {
     // do all the fancy stripe stuff...
-    event.preventDefault();
+
+    e.preventDefault(); // prevent the page from refreshing when the form is submitted
     setProcessing(true);
 
+    // Create a payment intent
     const payload = await stripe
       .confirmCardPayment(clientSecret, {
         payment_method: {
-          card: elements.getElement(CardElement),
+          card: elements.getElement(CardElement), // get the card details
         },
       })
       .then(({ paymentIntent }) => {
@@ -63,50 +65,45 @@ function Payment() {
             created: paymentIntent.created,
           });
 
-        setSucceeded(true);
-        setError(null);
-        setProcessing(false);
+        setSucceeded(true); // payment has been processed
+        setError(null); // no errors
+        setProcessing(false); // no longer processing
 
         dispatch({
           type: "EMPTY_BASKET",
         });
-
-        history.replace("/orders");
+        Navigate("/orders");
       });
   };
 
-  const handleChange = (event) => {
+  const handleChange = (e) => {
     // Listen for changes in the CardElement
     // and display any errors as the customer types their card details
-    setDisabled(event.empty);
-    setError(event.error ? event.error.message : "");
+    setDisabled(e.empty);
+    setError(e.error ? e.error.message : "");
   };
 
   return (
     <div className="payment">
-      <div className="payment-container">
+      <div className="payment_container">
         <h1>
           Checkout (<Link to="/checkout">{basket?.length} items</Link>)
         </h1>
-
-        {/* Payment section- delivery address */}
-        <div className="payment-section">
-          <div className="payment-title">
+        <div className="payment_section">
+          <div className="payment_title">
             <h3>Delivery Address</h3>
           </div>
-          <div className="payment-address">
+          <div className="payment_address">
             <p>{user?.email}</p>
             <p>123 React Lane</p>
             <p>Los Angeles, CA</p>
           </div>
         </div>
-
-        {/* Payment Section- Review items */}
-        <div className="payment-section">
-          <div className="payment-title">
+        <div className="payment_section">
+          <div className="payment_title">
             <h3>Review items and delivery</h3>
           </div>
-          <div className="payment-items">
+          <div className="payment_items">
             {basket.map((item) => (
               <CheckoutProduct
                 id={item.id}
@@ -118,19 +115,18 @@ function Payment() {
             ))}
           </div>
         </div>
-
-        {/* Payment section */}
-        <div className="payment-section">
-          <div className="payment-title">
+        <div className="payment_section">
+          <div className="payment_title">
             <h3>Payment Method</h3>
           </div>
-          <div className="payment-details">
+          <div className="payment_details">
             {/* Stripe magic will go */}
-
             <form onSubmit={handleSubmit}>
-              <CardElement onChange={handleChange} />
-
-              <div className="payment-priceContainer">
+              <CardElement
+                className="cardnumber_holder"
+                onChange={handleChange}
+              />
+              <div className="payment_priceContainer">
                 <CurrencyFormat
                   renderText={(value) => <h3>Order Total: {value}</h3>}
                   decimalScale={2}
@@ -144,7 +140,6 @@ function Payment() {
                 </button>
               </div>
 
-              {/* Errors */}
               {error && <div>{error}</div>}
             </form>
           </div>
